refactor : Participant를 인터페이스에서 추상클래스로 변경
