+++ src/main/java/domain/participant/Name.java
@@ -1,4 +1,4 @@
-package domain.player;
+package domain.participant;
 
 import java.util.Objects;
 
@@ -20,19 +20,6 @@ public class Name {
 
     public String getValue() {
         return value;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Name name = (Name) o;
-        return value.equals(name.value);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(value);
     }
 
     @Override
