+++ src/main/java/domain/participant/Player.java
@@ -1,0 +1,23 @@
+package domain.participant;
+
+import domain.card.Cards;
+
+public class Player extends Participant {
+
+    public Player(Name name, Cards cards) {
+        super(name, cards);
+    }
+
+    @Override
+    public boolean isFinished() {
+        return cards.isBust() || cards.isBlackJack();
+    }
+
+    @Override
+    public String toString() {
+        return "Player{" +
+                "name=" + name +
+                ", cards=" + cards +
+                '}';
+    }
+}
