--- src/main/java/domain/player/Participant.java
@@ -1,8 +1,0 @@
-package domain.player;
-
-import domain.card.Card;
-
-public interface Participant {
-    boolean isFinished();
-    void drawCard(Card card);
-}
