+++ src/test/java/domain/game/GameResultTest.java
@@ -4,9 +4,9 @@ import domain.card.Card;
 import domain.card.Cards;
 import domain.card.Denomination;
 import domain.card.Suit;
-import domain.player.Dealer;
-import domain.player.Name;
-import domain.player.Player;
+import domain.participant.Dealer;
+import domain.participant.Name;
+import domain.participant.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -33,10 +33,10 @@ class GameResultTest {
         return Stream.of(
                 Arguments.arguments("플레이어가 버스트가 아니고 딜러가 버스트인 경우",
                         new Player(new Name("abc"), new Cards(getCardList(Denomination.ACE, Denomination.NINE))),
-                        new Dealer(new Cards(getCardList(Denomination.QUEEN, Denomination.KING, Denomination.JACK)))),
+                        new Dealer(new Name("딜러"), new Cards(getCardList(Denomination.QUEEN, Denomination.KING, Denomination.JACK)))),
                 Arguments.arguments("둘 다 버스트가 아니고 딜러보다 숫자가 높은 경우",
                         new Player(new Name("abc"), new Cards(getCardList(Denomination.KING, Denomination.QUEEN))),
-                        new Dealer(new Cards(getCardList(Denomination.QUEEN, Denomination.NINE))))
+                        new Dealer(new Name("딜러"), new Cards(getCardList(Denomination.QUEEN, Denomination.NINE))))
         );
     }
 
@@ -46,7 +46,7 @@ class GameResultTest {
         Player player1 = new Player(new Name("abc1"), new Cards(getCardList(Denomination.ACE, Denomination.NINE)));      // 20
         Player player2 = new Player(new Name("abc2"), new Cards(getCardList(Denomination.EIGHT, Denomination.NINE)));    // 17
         Player player3 = new Player(new Name("abc3"), new Cards(getCardList(Denomination.ACE, Denomination.NINE)));      // 20
-        Dealer dealer = new Dealer(new Cards(getCardList(Denomination.QUEEN, Denomination.NINE)));                  // 19
+        Dealer dealer = new Dealer(new Name("딜러"), new Cards(getCardList(Denomination.QUEEN, Denomination.NINE)));                  // 19
 
         List<Player> players = Arrays.asList(player1, player2, player3);
         GameResult gameResult = new GameResult(players, dealer);
@@ -60,7 +60,7 @@ class GameResultTest {
         Player player1 = new Player(new Name("abc1"), new Cards(getCardList(Denomination.ACE, Denomination.NINE)));      // 20
         Player player2 = new Player(new Name("abc2"), new Cards(getCardList(Denomination.EIGHT, Denomination.NINE)));    // 17
         Player player3 = new Player(new Name("abc3"), new Cards(getCardList(Denomination.ACE, Denomination.NINE)));      // 20
-        Dealer dealer = new Dealer(new Cards(getCardList(Denomination.QUEEN, Denomination.NINE)));                  // 19
+        Dealer dealer = new Dealer(new Name("딜러"), new Cards(getCardList(Denomination.QUEEN, Denomination.NINE)));                  // 19
 
         List<Player> players = Arrays.asList(player1, player2, player3);
         GameResult gameResult = new GameResult(players, dealer);
