+++ src/test/java/domain/participant/DealerTest.java
@@ -1,4 +1,4 @@
-package domain.player;
+package domain.participant;
 
 import domain.card.Card;
 import domain.card.Cards;
@@ -16,20 +16,20 @@ import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class PlayerTest {
-
+public class DealerTest {
     @ParameterizedTest(name = "{0}")
     @MethodSource("provideParameters")
     @DisplayName("턴 강제 종료 여부")
-    void player(String comment,Cards cards, boolean expect) {
-        Player player = new Player(new Name("name"), cards);
-        assertThat(player.isFinished()).isEqualTo(expect);
+    void Dealer(String comment, Cards cards, boolean expect) {
+        Dealer dealer = new Dealer(new Name("딜러"), cards);
+        assertThat(dealer.isFinished()).isEqualTo(expect);
     }
 
     private static Stream<Arguments> provideParameters() {
         return Stream.of(
                 Arguments.arguments("합계 22인 경우 true", new Cards(getCardList(Denomination.TWO, Denomination.QUEEN, Denomination.KING)), true),
-                Arguments.arguments("합계 20인 경우 false", new Cards(getCardList(Denomination.QUEEN, Denomination.KING)), false)
+                Arguments.arguments("합계 17인 경우 true", new Cards(getCardList(Denomination.SEVEN, Denomination.QUEEN)), true),
+                Arguments.arguments("합계 15인 경우 false", new Cards(getCardList(Denomination.QUEEN, Denomination.FIVE)), false)
         );
     }
 
@@ -43,8 +43,8 @@ public class PlayerTest {
 
     @Test
     void drawCard() {
-        Player player = new Player(new Name("name"), new Cards(getCardList(Denomination.QUEEN)));
-        player.drawCard(new Card(Denomination.ACE, Suit.CLOVER));
-        assertThat(player.getCards().getValue().size()).isEqualTo(2);
+        Dealer dealer = new Dealer(new Name("딜러"), new Cards(getCardList(Denomination.QUEEN)));
+        dealer.drawCard(new Card(Denomination.ACE, Suit.CLOVER));
+        assertThat(dealer.getCards().getValue().size()).isEqualTo(2);
     }
 }
